Updated scheme.yml
Removed duplicates and organized alphabetically.  Modified Runs for most URL's to Safari since it's URL activated.  If it says Notification Center then I'm assuming that's correct.  It takes more time to test those and I don't have the time to check those.  I'd only say true on Jailbreak if it's a scheme that will only work when jailbroken.